Rename meta value file to match type
