+++ ironfish/src/wallet/database/metaValue.test.ts
@@ -1,7 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { MetaValue, MetaValueEncoding } from './meta'
+import { MetaValue, MetaValueEncoding } from './metaValue'
 
 describe('MetaValueEncoding', () => {
   describe('with a null value', () => {
